升级数据库插件 升级gradle到7.2.0

+++ app/src/main/java/com/web/misc/imageDraw/WaveDraw.kt
@@ -32,7 +32,7 @@ class WaveDraw:ImageDraw {
                     }
 
                 }
-                field[i/jump]= Math.hypot(v1, v2).toByte()
+                field[i/jump]= Math.hypot(v1, v2).toInt().toByte()
             }
             v?.invalidate()
         }
